third_party/blink/web_tests/virtual/view-transition-wide-gamut/external/wpt/css/css-view-transitions/pseudo-get-computed-style-clean-style-expected.txt

This is a testharness.js-based test.
[FAIL] properties of pseudo elements in update callback
  assert_equals: ::view-transition expected "absolute" but got "fixed"
Harness: the test ran to completion.
